fix: Resolve time.h and DWORD format compilation errors
- Add time.h include to settings.h for time functions
- Add localtime_s and fopen_s compatibility macros for non-MSVC compilers
- Fix DWORD format specifiers from %d to %lu with explicit casts in:
  * tray_ui.c: TrayUI_UpdateIcon and TrayUI_ShowContextMenu functions
  * config.c: Settings_Save function
- Ensures cross-compiler compatibility for time functions and format strings

+++ config.c
@@ -125,7 +125,7 @@ void Settings_Save(void) {
     sprintf_s(buffer, sizeof(buffer), "%d", g_stabilizer.enabled ? 1 : 0);
     WritePrivateProfileString("Settings", "Enabled", buffer, config_path);
     
-    sprintf_s(buffer, sizeof(buffer), "%d", g_stabilizer.delay_start_ms);
+    sprintf_s(buffer, sizeof(buffer), "%lu", (unsigned long)g_stabilizer.delay_start_ms);
     WritePrivateProfileString("Settings", "DelayStartMs", buffer, config_path);
     
     sprintf_s(buffer, sizeof(buffer), "%d", (int)(g_stabilizer.target_show_distance * 10));
